ta_support_project/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.6
djangorestframework==3.15.1
filelock==3.14.0
gunicorn==22.0.0
mysqlclient==2.2.4
packaging==24.0
platformdirs==4.2.2
sqlparse==0.5.0
virtualenv==20.26.2
